pyproject.toml
[tool.poetry]
name = "reforge-python"
version = "0.12.0"
description = "Python client for Reforge Feature Flags, Dynamic log levels, and Config as a Service: https://www.reforge.com"
license = "MIT"
authors = ["Michael Berkowitz <michael.berkowitz@gmail.com>", "James Kebinger <james.kebinger@reforge.com>"]
maintainers = ["Michael Berkowitz <michael.berkowitz@gmail.com>", "James Kebinger <james.kebinger@reforge.com>"]
readme = "README.md"
homepage = "https://www.reforge.com"
repository = "https://github.com/ReforgeHQ/sdk-python"
documentation = "https://docs.reforge.com/docs/sdks/python"
packages = [{include = "reforge_python"}, {include = "reforge_pb2.py"}]

[tool.poetry.dependencies]
cryptography = ">= 42.0.0"
python = ">= 3.9, < 4"
pyyaml = "^6.0.0"
mmh3 = ">=3.0.0,<5.0.0"
requests = ">= 2.30.0"
structlog = ">= 21.1.0"
sseclient-py = "^1.7.2"
cachetools = "^5.3.0"
protobuf = ">= 4.21.0"
isodate = "^0.6.1"
tenacity = ">=8.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
timecop = "^0.5.0dev"
mypy = "^1.4.1"
pre-commit = "^3.5.0"
responses = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
